+++ packages/gasket-plugin-elastic-apm/lib/index.d.ts
@@ -1,3 +1,4 @@
+import type { IncomingMessage, ServerResponse } from 'http';
 import type { AgentConfigOptions } from 'elastic-apm-node';
 
 declare module '@gasket/engine' {
@@ -5,5 +6,21 @@ declare module '@gasket/engine' {
     elasticAPM?: AgentConfigOptions & {
       sensitiveCookies?: Array<string>
     },
+  }
+
+  type RequestDetails = {
+    req: IncomingMessage,
+    res: ServerResponse
+  }
+
+  export interface HookExecTypes {
+    transactionName(
+      currentName: string,
+      details: RequestDetails
+    ): MaybeAsync<string>,
+    transactionLabels(
+      currentLabels: Record<string, string>,
+      details: RequestDetails
+    ): MaybeAsync<Record<string, string>>,
   }
 }