+++ packages/gasket-plugin-intl/lib/transaction-labels.js
@@ -1,0 +1,6 @@
+module.exports = (gasket, labels, { res }) => {
+  return {
+    ...labels,
+    locale: res.locals.gasketData.locale
+  };
+};