+++ packages/gasket-plugin-nextjs/lib/index.d.ts
@@ -1,4 +1,5 @@
 import type { WebpackContext } from '@gasket/plugin-webpack';
+import type { IncomingMessage, ServerResponse } from 'http'
 import type { NextConfig } from 'next/dist/next-server/server/config-shared';
 import type NextServer from 'next/dist/next-server/server/next-server';
 import type { MaybeAsync } from '@gasket/engine';
@@ -17,6 +18,11 @@ declare module '@gasket/engine' {
     nextExpress(params: {
       next: NextServer,
       express: Application
+    }): MaybeAsync<void>,
+    nextPreHandling(params: {
+      next: NextServer,
+      req: IncomingMessage,
+      res: ServerResponse
     }): MaybeAsync<void>
   }
 }
@@ -24,5 +30,17 @@ declare module '@gasket/engine' {
 declare module '@gasket/plugin-webpack' {
   export interface WebpackContext {
     isServer: boolean;
+  }
+}
+
+
+declare module 'http' {
+  export interface IncomingMessage {
+    getNextRoute(): Promise<null | {
+      page: string,
+      regex: RegExp,
+      routeKeys: Record<string, string>,
+      namedRegex: RegExp
+    }>
   }
 }