+++ packages/gasket-plugin-nextjs/lib/index.js
@@ -3,6 +3,9 @@ const url = require('url');
 const { name, devDependencies } = require('../package');
 const { createConfig } = require('./config');
 const { pluginIdentifier } = require('@gasket/resolve');
+const getNextRoute = require('./next-route');
+const transactionName = require('./transaction-name');
+const transactionLabels = require('./transaction-labels');
 
 const isDefined = (o) => typeof o !== 'undefined';
 
@@ -92,6 +95,17 @@ module.exports = {
         }
       }
     },
+    middleware: {
+      timing: {
+        before: ['@gasket/plugin-elastic-apm']
+      },
+      handler: (gasket) => [
+        (req, res, next) => {
+          req.getNextRoute = () => getNextRoute(gasket, req);
+          next();
+        }
+      ]
+    },
     express: {
       timing: {
         last: true
@@ -138,11 +152,21 @@ module.exports = {
         // interact with the express router we want to add a last, catch all
         // route that will activate the `next`.
         //
-        expressApp.all('*', app.getRequestHandler());
+        const nextHandler = app.getRequestHandler();
+        expressApp.all('*', async (req, res, next) => {
+          try {
+            await gasket.exec('nextPreHandling', { req, res, next: app });
+            nextHandler(req, res);
+          } catch (err) {
+            return void next(err);
+          }
+        });
 
         return app;
       }
     },
+    transactionName,
+    transactionLabels,
     build: async function build(gasket) {
       const { command } = gasket;
       // Don't do a build, use dev server for local
@@ -219,6 +243,13 @@ module.exports = {
           link: 'README.md#nextExpress',
           parent: 'express',
           after: 'next'
+        }, {
+          name: 'nextPreHandling',
+          method: 'exec',
+          description: 'Perform tasks just before next.js request handling',
+          link: 'README.md#nextPreHandling',
+          parent: 'express',
+          after: 'nextExpress'
         }],
         structures: [{
           name: 'pages/',